backend/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-CORS==3.0.10
mysql-connector-python==8.0.33
Flask-SQLAlchemy==2.5.1
GitPython==3.1.32
